requirements.txt
celery==5.0.2
Django==3.1.4
django-cors-headers==3.5.0
djangorestframework==3.12.1
gunicorn==20.0.4
psycopg2==2.8.6
